Fix: Do not format diff section
When using [the verbose option of git commit][1], the commit message
contains a unified diff at the end. That diff is currently also
formatted as if it is part of the commit message. Correct is to not
format it at all.

Git itself detects the diff by looking for a comment with the cut line
`------------------------ >8 ------------------------`. There might be
some more comment lines below that. But everything after that is
considered by git as the diff.

[1]: https://git-scm.com/docs/git-commit#Documentation/git-commit.txt--v