Preserve format of trailers
Git trailers are currently treated as flowing text paragraphs. This does
not work if there are more than one trailer.

This change adds a new chunk kind, Trailers. The last Paragraph chunk
(trailers must always be at the end of a commit message) of the
document is evaluated and if all lines of the paragraph are trailers, we
change the kind to Trailers and preserve the format of the paragraph.

This fixes #8